* fixed a printf() cast warning
git-svn-id: https://svn.code.sf.net/p/bzflag/code/branches/experimental/v2_99continuing/bzflag@21623 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ other_src/lua/src/lpeg.cpp
@@ -451,7 +451,7 @@ static const char *match (lua_State *L,
   for (;;) {
 #if defined(DEBUG)
       printf("s: |%s| stck: %d c: %d  ",
-             s, stack - getstackbase(L, ptop), captop);
+             s, (int)(stack - getstackbase(L, ptop)), captop);
       printinst(op, p);
 #endif
     switch ((Opcode)p->i.code) {
